Create Keras  important syntax
This file gives you a quick introduction to Keras for creating Artificial Neural Networks. It will also give you a good essence of Tensorflow 2.0. Keras is the offical API for Tensorflow 2.0. A quick glace through the file will give you enough Knowledge to start with Tensorflow 2.0.

I am very grateful to Jose Portilla to teach all the essential syntax of keras as a part of Complete Tenserflow 2.0 course on Udemy.